constructor of CoordinateTransformMesh changed to accept double width and height just as its parent TransformMesh

+++ mpicbg/src/main/java/mpicbg/models/CoordinateTransformMesh.java
@@ -29,8 +29,8 @@ public class CoordinateTransformMesh extends TransformMesh
 	public CoordinateTransformMesh(
 			final CoordinateTransform t,
 			final int numX,
-			final float width,
-			final float height )
+			final double width,
+			final double height )
 	{
 		super( numX, numY( numX, width, height ), width, height );
